// rs-lang/src/app/pages/textbook/models/user-aggregated-word.model.ts
import { UserWord } from "./user-word.model";

export interface UserAggregatedWord {
  _id?: string;
  group: number;
  page: number;
  word: string;
  image: string;
  audio: string;
  audioMeaning: string;
  audioExample: string;
  textMeaning: string;
  textExample: string;
  transcription: string;
  wordTranslate: string;
  textMeaningTranslate: string;
  textExampleTranslate: string;
  userWord?: UserWord;
}